% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zemi_moodle.R
\name{moodle_shinsei}
\alias{moodle_shinsei}
\title{Link page resources on Moodle site and faculty ID}
\usage{
moodle_shinsei(
  zemi_html = NULL,
  faculty_xlsx = NULL,
  name_course = "",
  project_id = ""
)
}
\arguments{
\item{zemi_html}{character. Path to the HTML file.}

\item{faculty_xlsx}{character. Path to the faculty xlsx file.}

\item{name_course}{character. Course name.}

\item{project_id}{character. Project id.}
}
\value{
matched result as tibble.
}
\description{
Link page resources on Moodle site and faculty ID
}
