% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omukyomu-package.R
\docType{package}
\name{omukyomu-package}
\alias{omukyomu}
\alias{omukyomu-package}
\title{omukyomu: Helper functions for Admin Tasks}
\description{
This package helps someone who doesn't use R to use R scripts for admin tasks.
}
\author{
\strong{Maintainer}: Kenji Sato \email{mail@kenjisato.jp}

}
\keyword{internal}
