% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_create}
\alias{project_create}
\title{Create kyomu project folder}
\usage{
project_create(name = "econkyomu", parent_dir = NULL)
}
\arguments{
\item{name}{character. Name of the directory}

\item{parent_dir}{character. Name of the parent directory.}
}
\value{
logical. TRUE when success.
}
\description{
Create kyomu project folder
}
